+++ crdesigner/ui/cli/command_line.py
@@ -6,6 +6,7 @@ from commonroad.planning.planning_problem import PlanningProblemSet
 from commonroad.common.file_writer import CommonRoadFileWriter, OverwriteExistingFile
 from commonroad.scenario.scenario import Tag
 
+from crdesigner.config.config import Lanelet2ConversionParams
 from crdesigner.start_gui import start_gui_new as start_gui
 from crdesigner.map_conversion.map_conversion_interface import commonroad_to_lanelet, lanelet_to_commonroad, \
     opendrive_to_commonroad, osm_to_commonroad, commonroad_to_sumo, sumo_to_commonroad
@@ -93,7 +94,14 @@ def main():
 
         if args.source_commonroad:
             if args.mode == "map-convert-lanelet":
-                commonroad_to_lanelet(input_file, output_file, args.proj, args.autoware, args.local_coordinates)
+                config = Lanelet2ConversionParams()
+                if args.proj:
+                    config.proj_string = args.proj
+                if args.autoware:
+                    config.autoware = args.autoware
+                if args.local_coordinates:
+                    config.local_coordinates = args.local_coordinates
+                commonroad_to_lanelet(input_file, output_file, config)
             if args.mode == "map-convert-sumo":
                 commonroad_to_sumo(input_file, output_file)
         else:
@@ -104,7 +112,14 @@ def main():
             elif args.mode == "map-convert-sumo":
                 scenario = sumo_to_commonroad(input_file)
             elif args.mode == "map-convert-lanelet":
-                scenario = lanelet_to_commonroad(input_file, args.proj, args.left_driving, args.adjacencies)
+                config = Lanelet2ConversionParams()
+                if args.proj:
+                    config.proj_string = args.proj
+                if args.adjacencies:
+                    config.adjacencies = args.adjacencies
+                if args.left_driving:
+                    config.left_driving = args.left_driving
+                scenario = lanelet_to_commonroad(input_file, config)
             else:
                 return
             tags = set([Tag(t) for t in args.tags])