+++ tests/map_conversion/lanelet2/test_lanelet2_to_cr_conversion.py
@@ -40,7 +40,7 @@ class TestLanelet2ToCommonRoadConversion(unittest.TestCase):
                   + f"/../test_maps/lanelet2/{osm_file_name}.osm", "r", ) as fh:
             osm = Lanelet2Parser(etree.parse(fh).getroot()).parse()
 
-        osm2l = Lanelet2CRConverter(proj_string="+proj=utm +zone=32 +ellps=WGS84")
+        osm2l = Lanelet2CRConverter()
         return osm2l(osm)
 
     def compare_maps(self, file_name: str) -> bool: