+++ tutorials/conversion_examples/example_opendrive_to_commonroad.py
@@ -1,19 +1,15 @@
 import os
-
 from commonroad.scenario.scenario import Tag
 from commonroad.common.file_writer import CommonRoadFileWriter, OverwriteExistingFile
 from commonroad.planning.planning_problem import PlanningProblemSet
-
 from crdesigner.map_conversion.opendrive.opendrive_parser.parser import parse_opendrive
 from crdesigner.map_conversion.opendrive.opendrive_conversion.network import Network
-
 from crdesigner.map_conversion.map_conversion_interface import opendrive_to_commonroad
-from crdesigner.configurations.get_configs import get_configs
+from crdesigner.config.config import OpenDRIVEConversionParams
 
 input_path = ""  # replace empty string
-
-# load configuration
-config = get_configs()
+config = OpenDRIVEConversionParams()
+config.lanelet_types_backwards_compatible = False
 
 # ----------------------------------------------- Option 1: General API ------------------------------------------------
 # load OpenDRIVE file, parse it, and convert it to a CommonRoad scenario
@@ -24,9 +20,9 @@ scenario = opendrive_to_commonroad(input_path)
 writer = CommonRoadFileWriter(
     scenario=scenario,
     planning_problem_set=PlanningProblemSet(),
-    author=config.file_header.author,
-    affiliation=config.file_header.affiliation,
-    source=config.file_header.source,
+    author="Sebastian Maierhofer",
+    affiliation="Technical University of Munich",
+    source="CommonRoad Scenario Designer",
     tags={Tag.URBAN},
 )
 writer.write_to_file(os.path.dirname(os.path.realpath(__file__)) + "/" + "ZAM_OpenDRIVETest-1_1-T1.xml",
@@ -38,7 +34,7 @@ writer.write_to_file(os.path.dirname(os.path.realpath(__file__)) + "/" + "ZAM_Op
 opendrive = parse_opendrive(input_path)
 
 # create OpenDRIVE intermediate network object from configuration
-road_network = Network(config.opendrive)
+road_network = Network(config)
 
 # convert OpenDRIVE file
 road_network.load_opendrive(opendrive)
@@ -50,9 +46,9 @@ scenario = road_network.export_commonroad_scenario()
 writer = CommonRoadFileWriter(
     scenario=scenario,
     planning_problem_set=PlanningProblemSet(),
-    author=config.file_header.author,
-    affiliation=config.file_header.affiliation,
-    source=config.file_header.source,
+    author="Sebastian Maierhofer",
+    affiliation="Technical University of Munich",
+    source="CommonRoad Scenario Designer",
     tags={Tag.URBAN},
 )
 writer.write_to_file(os.path.dirname(os.path.realpath(__file__)) + "/" + "ZAM_OpenDRIVETest-1_1-T2.xml",